# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Oru'el GPU Relay API
  version: "1.0.0"
  description: |
    Oru'el's relay to the Spheron AI GPU network. Every endpoint here mirrors the
    shape of Spheron's own API (https://docs.spheron.ai/api-reference) field-for-field,
    so an integration built against Spheron's docs needs only two changes to work here:

    1. Point the base URL at Oru'el's relay instead of `https://app.spheron.ai`.
    2. Authenticate with `X-API-Key: <key Oru'el issued you>` instead of a Spheron
       bearer token.

    **Pricing note:** every dollar figure returned by this API (offer prices, hourly
    rates, total cost, volume storage rates, etc.) already includes Oru'el's markup.
    There is nothing further to add on your side - the number you see is the number
    you're billed.

    **Scoping note:** `GET /deployments` and `GET /deployments/{id}` only ever return
    deployments launched with *your* API key. You cannot see or affect another
    partner's instances.

    Endpoints intentionally not exposed here (and why): `GET /balance` and
    `GET /teams` from Spheron's own API expose Oru'el's internal account balance and
    team membership - that's not relayed to any downstream partner.
  contact:
    name: Oru'el
...
servers:
  - url: https://relay.oruel.example/api
    description: Production (replace with your real deployed hostname)
  - url: http://localhost:4000/api
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /volumes:
    get:
      tags: [Volumes]
      summary: List persistent storage volumes
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: status, in: query, schema: { type: string, enum: [available, attached, deleting, deleted] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VolumePage" }
    post:
      tags: [Volumes]
      summary: Create a volume
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVolumeRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
  /volumes/pricing:
    get:
      tags: [Volumes]
      summary: Get storage pricing per GB per hour
      parameters:
        - { name: provider, in: query, schema: { type: string, enum: [voltage-park, data-crunch, sesterce, spheron-ai] } }
        - { name: region, in: query, schema: { type: string } }
        - { name: cloudId, in: query, schema: { type: string }, description: "Sesterce only" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider: { type: string }
                  hourlyRatePerGb: { type: number, format: float }
  /volumes/regions:
    get:
      tags: [Volumes]
      summary: List valid regions (and tiers/pricing) for a storage provider
      parameters:
        - { name: provider, in: query, schema: { type: string, enum: [voltage-park, data-crunch, sesterce, spheron-ai], default: voltage-park } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasAnyOffers: { type: boolean }
                  regions:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        hasGpuOffers: { type: boolean }
                        minSizeGb: { type: integer }
                        maxSizeGb: { type: integer }
                        hourlyRatePerGb: { type: number, format: float }
  /volumes/{volumeId}:
    parameters:
      - { name: volumeId, in: path, required: true, schema: { type: string } }
    get:
      tags: [Volumes]
      summary: Get volume details
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
    patch:
      tags: [Volumes]
      summary: Update a volume (rename/resize where the provider allows it)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                sizeInGb: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
    delete:
      tags: [Volumes]
      summary: Delete a volume
      description: Volume must not be attached to any active deployment.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { success: { type: boolean } } }
  /volumes/{volumeId}/attach:
    parameters:
      - { name: volumeId, in: path, required: true, schema: { type: string } }
    post:
      tags: [Volumes]
      summary: Attach a volume to a running deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deploymentId]
              properties: { deploymentId: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
  /volumes/{volumeId}/detach:
    parameters:
      - { name: volumeId, in: path, required: true, schema: { type: string } }
    post:
      tags: [Volumes]
      summary: Detach a volume from a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deploymentId]
              properties: { deploymentId: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volume" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: The key Oru'el issued you. Never share it or embed it in client-side code.
...
  schemas:
...
    VolumePage:
      type: object
      properties:
        volumes:
          type: array
          items: { $ref: "#/components/schemas/Volume" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    Volume:
      type: object
      properties:
        id: { type: string }
        volumeId: { type: string }
        name: { type: string }
        providerId: { type: string }
        sizeInGb: { type: integer }
        usedCapacityBytes: { type: integer }
        status: { type: string, enum: [available, attached, deleting, deleted] }
        hourlyRate: { type: number, format: float, description: "USD/GB/hr, markup already applied" }
        attachedToDeploymentIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    CreateVolumeRequest:
      type: object
      required: [name, sizeInGb, provider]
      properties:
        name: { type: string, description: "lowercase alphanumeric with hyphens/underscores, max 60 chars" }
        sizeInGb: { type: integer }
        provider: { type: string, enum: [voltage-park, data-crunch, sesterce, spheron-ai] }
        region: { type: string, description: "Required for data-crunch, sesterce, spheron-ai" }
        cloudId: { type: string, description: "Required for sesterce only" }
        deploymentId: { type: string, description: "Optionally attach at creation time" }